tserver tofu: only create fping6 link when needed

+++ t_server/tofu/modules/tserver_user_data/provision/install-openvpn-build-deps-rhel-9.sh
@@ -53,7 +53,7 @@ liboping \
 libdb-utils
 
 # Make sure that fping6 can be found
-ln -s /usr/sbin/fping /usr/sbin/fping6
+test -e /usr/sbin/fping6 || ln -s /usr/sbin/fping /usr/sbin/fping6
 
 # Old OpenVPN version support with podman
 dnf -y install \
